refactor: separate character models
